# openapi.yaml
openapi: 3.0.0
info:
  title: ConnectVerse API
  description: API for ConnectVerse, a unified messaging platform for modern teams.
  version: 1.0.0
servers:
  - url: /api
    description: Local API server
paths:
...
  /emails:
    get:
      summary: List Emails
      description: Get emails from a specific folder.
      tags: [Emails]
      parameters:
        - name: folder
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/EmailFolder'
      responses:
        '200':
          description: A list of emails.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Email'
    post:
      summary: Send Email
      description: Compose and send a new email.
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEmail'
      responses:
        '201':
          description: Email sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Email'
  /emails/{emailId}:
    put:
      summary: Update Email
      description: Update an email's properties (e.g., mark as read).
      tags: [Emails]
      parameters:
        - name: emailId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isRead:
                  type: boolean
      responses:
        '200':
          description: Email updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Email'
...
components:
  schemas:
...
    EmailFolder:
      type: string
      enum: [inbox, sent, drafts, trash]
    Email:
      type: object
      properties:
        id:
          type: string
        from:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
              format: email
            avatar:
              type: string
              format: uri
        subject:
          type: string
        body:
          type: string
        timestamp:
          type: string
          format: date-time
        isRead:
          type: boolean
        folder:
          $ref: '#/components/schemas/EmailFolder'
    NewEmail:
      type: object
      required: [subject, body]
      properties:
        subject:
          type: string
        body:
          type: string
...